refactor code javascript
