+++ NodeJS/server.js
@@ -9,14 +9,12 @@ io.on('connection', function (socket) {
     console.log("new client connected");
     var redisClient = redis.createClient();
     redisClient.subscribe('comment', 'votes', 'closePoll', 'pollTimeout', 'deleteParticipant', 'reopenPoll');
-
     redisClient.on("message", function(channel, message) {
-        console.log("new message in queue "+ message + "channel");
+        console.log("channel: " + channel);
         socket.emit(channel, message);
     });
 
     socket.on('disconnect', function() {
         redisClient.quit();
     });
-
 });
