+++ bower.json
@@ -18,12 +18,7 @@
   "dependencies": {
     "components-font-awesome": "^4.6.3",
     "bootstrap": "^3.3.7",
-    "bootstrap-toggle": "^2.2.2",
-    "jquery.steps": "^1.1.0",
     "bootstrap-tagsinput": "^0.8.0",
-    "datatables.net": "^1.10.12",
-    "datatables.net-bs": "^1.10.12",
-    "jQuery": "^3.1.1",
     "animate.css": "^3.5.2",
     "bootstrap-social": "^5.0.0",
     "eonasdan-bootstrap-datetimepicker": "^4.17.43",
