+++ resources/assets/js/vote.js
@@ -30,32 +30,32 @@ $(document).ready(function(){
             if (emailVote.trim() != '') {
                 if (testEmail.test(emailVote)) {
                     this.disabled = true;
-                    $('.message-validation').removeClass('alert alert-warning');
+                    $('.message-validation').removeClass('alert alert-warning alert-poll-set-ip');
                     $('.message-validation').html('');
                     $('#form-vote').submit();
                     $('.loader').show();
                 } else {
                     divChangeAmount = $(this).parent();
-                    $('.message-validation').addClass('alert alert-warning');
+                    $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                     var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' +  divChangeAmount.data('messageValidateEmail');
                     $('.message-validation').html(message);
                 }
             } else {
                 divChangeAmount = $(this).parent();
-                $('.message-validation').addClass('alert alert-warning');
+                $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                 var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' + divChangeAmount.data('messageRequiredEmail');
                 $('.message-validation').html(message);
             }
         } else if (isRequiredName == 1) {
             if (nameVote.trim() != '') {
                 this.disabled = true;
-                $('.message-validation').removeClass('alert alert-warning');
+                $('.message-validation').removeClass('alert alert-warning alert-poll-set-ip');
                 $('.message-validation').html('');
                 $('#form-vote').submit();
                 $('.loader').show();
             } else {
                 divChangeAmount = $(this).parent();
-                $('.message-validation').addClass('alert alert-warning');
+                $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                 var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' + divChangeAmount.data('messageRequiredName');
                 $('.message-validation').html(message);
             }
@@ -63,19 +63,19 @@ $(document).ready(function(){
             if (nameVote.trim() != '') {
                 if (emailVote.trim() == '') {
                     divChangeAmount = $(this).parent();
-                    $('.message-validation').addClass('alert alert-warning');
+                    $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                     var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' + divChangeAmount.data('messageRequiredEmail');
                     $('.message-validation').html(message);
                 } else {
                     if (testEmail.test(emailVote)) {
                         this.disabled = true;
-                        $('.message-validation').removeClass('alert alert-warning');
+                        $('.message-validation').removeClass('alert alert-warning alert-poll-set-ip');
                         $('.message-validation').html('');
                         $('#form-vote').submit();
                         $('.loader').show();
                     } else {
                         divChangeAmount = $(this).parent();
-                        $('.message-validation').addClass('alert alert-warning');
+                        $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                         var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' +  divChangeAmount.data('messageValidateEmail');
                         $('.message-validation').html(message);
                     }
@@ -83,12 +83,12 @@ $(document).ready(function(){
             } else {
                 if (emailVote.trim() == '') {
                     divChangeAmount = $(this).parent();
-                    $('.message-validation').addClass('alert alert-warning');
+                    $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                     var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' + divChangeAmount.data('messageRequiredNameAndEmail');
                     $('.message-validation').html(message);
                 } else {
                     divChangeAmount = $(this).parent();
-                    $('.message-validation').addClass('alert alert-warning');
+                    $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                     var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' + divChangeAmount.data('messageRequiredName');
                     $('.message-validation').html(message);
                 }
@@ -100,7 +100,7 @@ $(document).ready(function(){
                     isPassValidate = true;
                 } else {
                     divChangeAmount = $(this).parent();
-                    $('.message-validation').addClass('alert alert-warning');
+                    $('.message-validation').addClass('alert alert-warning alert-poll-set-ip');
                     var message = "<span class='glyphicon glyphicon-warning-sign'></span>" + ' ' +  divChangeAmount.data('messageValidateEmail');
                     $('.message-validation').html(message);
                 }
@@ -110,7 +110,7 @@ $(document).ready(function(){
 
             if (isPassValidate) {
                 this.disabled = true;
-                $('.message-validation').removeClass('alert alert-warning');
+                $('.message-validation').removeClass('alert alert-warning alert-poll-set-ip');
                 $('.message-validation').html('');
                 $('#form-vote').submit();
                 $('.loader').show();
