+++ resources/assets/sass/user.scss
@@ -267,3 +267,15 @@ li {
 .btn-social-custom {
     margin-bottom: 4px;
 }
+
+.wrap-text {
+    word-wrap: break-word;
+    white-space: normal;
+}
+
+.message-infor-detail {
+    background: white;
+    border-color: white;
+    border-radius: 0;
+    color: green;
+}
