+++ resources/lang/ja/history.php
