+++ resources/lang/ja/pagination.php
