+++ resources/lang/ja/polls.php
